+++ src/client/grid/TileCountOverlay.tsx
@@ -104,7 +104,11 @@ export default memo(function TileCountOverlay({ grid }: TileCountOverlayProps) {
       grid.height * tileSize + 2 * BLEED
     );
 
-    if (!position || Number.isNaN(position.x) || Number.isNaN(position.y))
+    if (
+      !position ||
+      !Number.isFinite(position.x) ||
+      !Number.isFinite(position.y)
+    )
       return;
     const tile = grid.getTile(position.x, position.y);
     if (!tile.exists) return;